+++ lib/libesp32/Berry/default/be_modtab.c
@@ -100,6 +100,10 @@ extern void be_load_driverlib(bvm *vm);
 extern void be_load_driver_i2c_lib(bvm *vm);
 extern void be_load_md5_lib(bvm *vm);
 
+#ifdef USE_I2S_AUDIO_BERRY
+extern void be_load_driver_audio_lib(bvm *vm);
+#endif
+
 #ifdef USE_LVGL
 extern void be_load_lvgl_color_lib(bvm *vm);
 extern void be_load_lvgl_font_lib(bvm *vm);
@@ -126,6 +130,9 @@ BERRY_API void be_load_custom_libs(bvm *vm)
     be_load_wirelib(vm);
     be_load_driver_i2c_lib(vm);
 #endif // USE_I2C
+#ifdef USE_I2S_AUDIO_BERRY
+    be_load_driver_audio_lib(vm);
+#endif
 #ifdef USE_LVGL
     // LVGL
     be_load_lvgl_color_lib(vm);