+++ lib/libesp32/Berry/default/be_tasmotalib.c
@@ -1368,193 +1368,200 @@ be_local_closure(event,   /* name */
     0,                          /* has sup protos */
     NULL,                       /* no sub protos */
     1,                          /* has constants */
-    ( &(const bvalue[29]) {     /* upvals */
-      { { .s=be_nested_const_str("cmd", -158181397, 3) }, BE_STRING},
-      { { .s=be_nested_const_str("exec_cmd", 493567399, 8) }, BE_STRING},
-      { { .s=be_nested_const_str("rule", -64077613, 4) }, BE_STRING},
-      { { .s=be_nested_const_str("exec_rules", 1445221092, 10) }, BE_STRING},
-      { { .s=be_nested_const_str("mqtt_data", -1756753932, 9) }, BE_STRING},
-      { { .s=be_nested_const_str("gc", 1042313471, 2) }, BE_STRING},
-      { { .s=be_nested_const_str("every_50ms", -1911083288, 10) }, BE_STRING},
-      { { .s=be_nested_const_str("run_deferred", 371594696, 12) }, BE_STRING},
-      { { .s=be_nested_const_str("_drivers", -1034638311, 8) }, BE_STRING},
-      { { .s=be_nested_const_str("every_second", 2075451465, 12) }, BE_STRING},
-      { { .s=be_nested_const_str("every_100ms", 1546407804, 11) }, BE_STRING},
-      { { .s=be_nested_const_str("web_add_button", -757092238, 14) }, BE_STRING},
-      { { .s=be_nested_const_str("web_add_main_button", -334599632, 19) }, BE_STRING},
-      { { .s=be_nested_const_str("web_add_management_button", -1556090110, 25) }, BE_STRING},
-      { { .s=be_nested_const_str("web_add_config_button", 639674325, 21) }, BE_STRING},
-      { { .s=be_nested_const_str("web_add_console_button", -813531104, 22) }, BE_STRING},
-      { { .s=be_nested_const_str("save_before_restart", 1253239338, 19) }, BE_STRING},
-      { { .s=be_nested_const_str("web_add_handler", -304792334, 15) }, BE_STRING},
-      { { .s=be_nested_const_str("web_sensor", -1394870324, 10) }, BE_STRING},
-      { { .s=be_nested_const_str("json_append", -1292948012, 11) }, BE_STRING},
-      { { .s=be_nested_const_str("button_pressed", 1694209616, 14) }, BE_STRING},
-      { { .s=be_nested_const_str("button_preselif", 491115394, 15) }, BE_STRING},
-      { { .s=be_nested_const_str("display", 1164572437, 7) }, BE_STRING},
-      { { .s=be_nested_const_str("string", 398550328, 6) }, BE_STRING},
-      { { .s=be_nested_const_str("log", 1062293841, 3) }, BE_STRING},
-      { { .s=be_nested_const_str("format", -1180859054, 6) }, BE_STRING},
-      { { .s=be_nested_const_str("BRY: exception %s - %m", -1290966132, 22) }, BE_STRING},
-      { { .i=3 }, BE_INT},
-      { { .s=be_nested_const_str("stop_iteration", -121173395, 14) }, BE_STRING},
+    ( &(const bvalue[26]) {     /* constants */
+      be_nested_string("every_50ms", -1911083288, 10),
+      be_nested_string("run_deferred", 371594696, 12),
+      be_nested_string("cmd", -158181397, 3),
+      be_nested_string("exec_cmd", 493567399, 8),
+      be_nested_string("rule", -64077613, 4),
+      be_nested_string("exec_rules", 1445221092, 10),
+      be_nested_string("mqtt_data", -1756753932, 9),
+      be_nested_string("gc", 1042313471, 2),
+      be_nested_string("_drivers", -1034638311, 8),
+      be_nested_string("every_second", 2075451465, 12),
+      be_nested_string("every_100ms", 1546407804, 11),
+      be_nested_string("web_add_button", -757092238, 14),
+      be_nested_string("web_add_main_button", -334599632, 19),
+      be_nested_string("web_add_management_button", -1556090110, 25),
+      be_nested_string("web_add_config_button", 639674325, 21),
+      be_nested_string("web_add_console_button", -813531104, 22),
+      be_nested_string("save_before_restart", 1253239338, 19),
+      be_nested_string("web_add_handler", -304792334, 15),
+      be_nested_string("web_sensor", -1394870324, 10),
+      be_nested_string("json_append", -1292948012, 11),
+      be_nested_string("button_pressed", 1694209616, 14),
+      be_nested_string("display", 1164572437, 7),
+      be_nested_string("string", 398550328, 6),
+      be_nested_string("format", -1180859054, 6),
+      be_nested_string("BRY: Exception> '%s' - %s", -2047976332, 25),
+      be_nested_string("stop_iteration", -121173395, 14),
     }),
     (be_nested_const_str("event", -30355297, 5)),
     (be_nested_const_str("input", -103256197, 5)),
-    ( &(const binstruction[153]) {  /* code */
+    ( &(const binstruction[163]) {  /* code */
       0x1C140300,  //  0000  EQ	R5	R1	R256
-      0x78160006,  //  0001  JMPF	R5	#0009
+      0x78160001,  //  0001  JMPF	R5	#0004
       0x8C140101,  //  0002  GETMET	R5	R0	R257
-      0x5C1C0400,  //  0003  MOVE	R7	R2
-      0x5C200600,  //  0004  MOVE	R8	R3
-      0x5C240800,  //  0005  MOVE	R9	R4
-      0x7C140800,  //  0006  CALL	R5	4
-      0x80040A00,  //  0007  RET	1	R5
-      0x7002008E,  //  0008  JMP		#0098
-      0x1C140302,  //  0009  EQ	R5	R1	R258
-      0x78160004,  //  000A  JMPF	R5	#0010
-      0x8C140103,  //  000B  GETMET	R5	R0	R259
-      0x5C1C0800,  //  000C  MOVE	R7	R4
-      0x7C140400,  //  000D  CALL	R5	2
-      0x80040A00,  //  000E  RET	1	R5
-      0x70020087,  //  000F  JMP		#0098
-      0x1C140304,  //  0010  EQ	R5	R1	R260
-      0x78160002,  //  0011  JMPF	R5	#0015
-      0x4C140000,  //  0012  LDNIL	5
-      0x80040A00,  //  0013  RET	1	R5
-      0x70020082,  //  0014  JMP		#0098
-      0x1C140305,  //  0015  EQ	R5	R1	R261
-      0x78160003,  //  0016  JMPF	R5	#001B
-      0x8C140105,  //  0017  GETMET	R5	R0	R261
-      0x7C140200,  //  0018  CALL	R5	1
-      0x80040A00,  //  0019  RET	1	R5
-      0x7002007C,  //  001A  JMP		#0098
-      0x1C140306,  //  001B  EQ	R5	R1	R262
-      0x78160003,  //  001C  JMPF	R5	#0021
-      0x8C140107,  //  001D  GETMET	R5	R0	R263
-      0x7C140200,  //  001E  CALL	R5	1
-      0x80040A00,  //  001F  RET	1	R5
-      0x70020076,  //  0020  JMP		#0098
-      0x88140108,  //  0021  GETMBR	R5	R0	R264
-      0x78160074,  //  0022  JMPF	R5	#0098
-      0x60140000,  //  0023  GETGBL	R5	G0
-      0x88180108,  //  0024  GETMBR	R6	R0	R264
-      0x7C140200,  //  0025  CALL	R5	1
-      0xA802006D,  //  0026  EXBLK	0	#0095
-      0x5C180A00,  //  0027  MOVE	R6	R5
-      0x7C180000,  //  0028  CALL	R6	0
-      0xA802005E,  //  0029  EXBLK	0	#0089
-      0x1C1C0309,  //  002A  EQ	R7	R1	R265
-      0x781E0004,  //  002B  JMPF	R7	#0031
-      0x881C0D09,  //  002C  GETMBR	R7	R6	R265
-      0x781E0002,  //  002D  JMPF	R7	#0031
-      0x8C1C0D09,  //  002E  GETMET	R7	R6	R265
-      0x7C1C0200,  //  002F  CALL	R7	1
-      0x70020055,  //  0030  JMP		#0087
-      0x1C1C030A,  //  0031  EQ	R7	R1	R266
-      0x781E0004,  //  0032  JMPF	R7	#0038
-      0x881C0D0A,  //  0033  GETMBR	R7	R6	R266
-      0x781E0002,  //  0034  JMPF	R7	#0038
-      0x8C1C0D0A,  //  0035  GETMET	R7	R6	R266
-      0x7C1C0200,  //  0036  CALL	R7	1
-      0x7002004E,  //  0037  JMP		#0087
-      0x1C1C030B,  //  0038  EQ	R7	R1	R267
-      0x781E0004,  //  0039  JMPF	R7	#003F
-      0x881C0D0B,  //  003A  GETMBR	R7	R6	R267
-      0x781E0002,  //  003B  JMPF	R7	#003F
-      0x8C1C0D0B,  //  003C  GETMET	R7	R6	R267
-      0x7C1C0200,  //  003D  CALL	R7	1
-      0x70020047,  //  003E  JMP		#0087
-      0x1C1C030C,  //  003F  EQ	R7	R1	R268
-      0x781E0004,  //  0040  JMPF	R7	#0046
-      0x881C0D0C,  //  0041  GETMBR	R7	R6	R268
-      0x781E0002,  //  0042  JMPF	R7	#0046
-      0x8C1C0D0C,  //  0043  GETMET	R7	R6	R268
-      0x7C1C0200,  //  0044  CALL	R7	1
-      0x70020040,  //  0045  JMP		#0087
-      0x1C1C030D,  //  0046  EQ	R7	R1	R269
-      0x781E0004,  //  0047  JMPF	R7	#004D
-      0x881C0D0D,  //  0048  GETMBR	R7	R6	R269
-      0x781E0002,  //  0049  JMPF	R7	#004D
-      0x8C1C0D0D,  //  004A  GETMET	R7	R6	R269
-      0x7C1C0200,  //  004B  CALL	R7	1
-      0x70020039,  //  004C  JMP		#0087
-      0x1C1C030E,  //  004D  EQ	R7	R1	R270
-      0x781E0004,  //  004E  JMPF	R7	#0054
-      0x881C0D0E,  //  004F  GETMBR	R7	R6	R270
-      0x781E0002,  //  0050  JMPF	R7	#0054
-      0x8C1C0D0E,  //  0051  GETMET	R7	R6	R270
-      0x7C1C0200,  //  0052  CALL	R7	1
-      0x70020032,  //  0053  JMP		#0087
-      0x1C1C030F,  //  0054  EQ	R7	R1	R271
-      0x781E0004,  //  0055  JMPF	R7	#005B
-      0x881C0D0F,  //  0056  GETMBR	R7	R6	R271
-      0x781E0002,  //  0057  JMPF	R7	#005B
-      0x8C1C0D0F,  //  0058  GETMET	R7	R6	R271
-      0x7C1C0200,  //  0059  CALL	R7	1
-      0x7002002B,  //  005A  JMP		#0087
-      0x1C1C0310,  //  005B  EQ	R7	R1	R272
-      0x781E0004,  //  005C  JMPF	R7	#0062
-      0x881C0D10,  //  005D  GETMBR	R7	R6	R272
-      0x781E0002,  //  005E  JMPF	R7	#0062
-      0x8C1C0D10,  //  005F  GETMET	R7	R6	R272
-      0x7C1C0200,  //  0060  CALL	R7	1
-      0x70020024,  //  0061  JMP		#0087
-      0x1C1C0311,  //  0062  EQ	R7	R1	R273
-      0x781E0004,  //  0063  JMPF	R7	#0069
-      0x881C0D11,  //  0064  GETMBR	R7	R6	R273
-      0x781E0002,  //  0065  JMPF	R7	#0069
-      0x8C1C0D11,  //  0066  GETMET	R7	R6	R273
-      0x7C1C0200,  //  0067  CALL	R7	1
-      0x7002001D,  //  0068  JMP		#0087
-      0x1C1C0312,  //  0069  EQ	R7	R1	R274
-      0x781E0004,  //  006A  JMPF	R7	#0070
-      0x881C0D12,  //  006B  GETMBR	R7	R6	R274
-      0x781E0002,  //  006C  JMPF	R7	#0070
-      0x8C1C0D12,  //  006D  GETMET	R7	R6	R274
-      0x7C1C0200,  //  006E  CALL	R7	1
-      0x70020016,  //  006F  JMP		#0087
-      0x1C1C0313,  //  0070  EQ	R7	R1	R275
-      0x781E0004,  //  0071  JMPF	R7	#0077
-      0x881C0D13,  //  0072  GETMBR	R7	R6	R275
-      0x781E0002,  //  0073  JMPF	R7	#0077
-      0x8C1C0D13,  //  0074  GETMET	R7	R6	R275
-      0x7C1C0200,  //  0075  CALL	R7	1
-      0x7002000F,  //  0076  JMP		#0087
-      0x1C1C0314,  //  0077  EQ	R7	R1	R276
-      0x781E0007,  //  0078  JMPF	R7	#0081
-      0x881C0D14,  //  0079  GETMBR	R7	R6	R276
-      0x781E0005,  //  007A  JMPF	R7	#0081
-      0x1C1C0311,  //  007B  EQ	R7	R1	R273
-      0x781DFFFE,  //  007C  JMPF	R7	#007C
-      0x881C0D16,  //  007D  GETMBR	R7	R6	R278
-      0x8C1C0D16,  //  007E  GETMET	R7	R6	R278
-      0x7C1C0200,  //  007F  CALL	R7	1
-      0x70020005,  //  0080  JMP		#0087
-      0x1C1C0316,  //  0081  EQ	R7	R1	R278
-      0x781E0003,  //  0082  JMPF	R7	#0087
-      0x881C0D16,  //  0083  GETMBR	R7	R6	R278
-      0x781E0001,  //  0084  JMPF	R7	#0087
-      0x8C1C0D16,  //  0085  GETMET	R7	R6	R278
-      0x7C1C0200,  //  0086  CALL	R7	1
-      0xA8040001,  //  0087  EXBLK	1	1
-      0x7002000A,  //  0088  JMP		#0094
-      0xAC1C0002,  //  0089  CATCH	R7	0	2
-      0x70020007,  //  008A  JMP		#0093
-      0xA4262E00,  //  008B  IMPORT	R9	R279
-      0x8C280118,  //  008C  GETMET	R10	R0	R280
-      0x8C301319,  //  008D  GETMET	R12	R9	R281
-      0x5838001A,  //  008E  LDCONST	R14	K26
-      0x583C001B,  //  008F  LDCONST	R15	K27
-      0x7C300600,  //  0090  CALL	R12	3
-      0x7C280400,  //  0091  CALL	R10	2
-      0x70020000,  //  0092  JMP		#0094
-      0xB0080000,  //  0093  RAISE	2	R0	R0
-      0x7001FF91,  //  0094  JMP		#0027
-      0x5814001C,  //  0095  LDCONST	R5	K28
-      0xAC140200,  //  0096  CATCH	R5	1	0
-      0xB0080000,  //  0097  RAISE	2	R0	R0
-      0x80000000,  //  0098  RET	0	R0
+      0x7C140200,  //  0003  CALL	R5	1
+      0x1C140302,  //  0004  EQ	R5	R1	R258
+      0x78160006,  //  0005  JMPF	R5	#000D
+      0x8C140103,  //  0006  GETMET	R5	R0	R259
+      0x5C1C0400,  //  0007  MOVE	R7	R2
+      0x5C200600,  //  0008  MOVE	R8	R3
+      0x5C240800,  //  0009  MOVE	R9	R4
+      0x7C140800,  //  000A  CALL	R5	4
+      0x80040A00,  //  000B  RET	1	R5
+      0x70020094,  //  000C  JMP		#00A2
+      0x1C140304,  //  000D  EQ	R5	R1	R260
+      0x78160004,  //  000E  JMPF	R5	#0014
+      0x8C140105,  //  000F  GETMET	R5	R0	R261
+      0x5C1C0800,  //  0010  MOVE	R7	R4
+      0x7C140400,  //  0011  CALL	R5	2
+      0x80040A00,  //  0012  RET	1	R5
+      0x7002008D,  //  0013  JMP		#00A2
+      0x1C140306,  //  0014  EQ	R5	R1	R262
+      0x78160002,  //  0015  JMPF	R5	#0019
+      0x4C140000,  //  0016  LDNIL	5
+      0x80040A00,  //  0017  RET	1	R5
+      0x70020088,  //  0018  JMP		#00A2
+      0x1C140307,  //  0019  EQ	R5	R1	R263
+      0x78160003,  //  001A  JMPF	R5	#001F
+      0x8C140107,  //  001B  GETMET	R5	R0	R263
+      0x7C140200,  //  001C  CALL	R5	1
+      0x80040A00,  //  001D  RET	1	R5
+      0x70020082,  //  001E  JMP		#00A2
+      0x88140108,  //  001F  GETMBR	R5	R0	R264
+      0x78160080,  //  0020  JMPF	R5	#00A2
+      0x60140000,  //  0021  GETGBL	R5	G0
+      0x88180108,  //  0022  GETMBR	R6	R0	R264
+      0x7C140200,  //  0023  CALL	R5	1
+      0xA8020079,  //  0024  EXBLK	0	#009F
+      0x5C180A00,  //  0025  MOVE	R6	R5
+      0x7C180000,  //  0026  CALL	R6	0
+      0xA8020069,  //  0027  EXBLK	0	#0092
+      0x1C1C0309,  //  0028  EQ	R7	R1	R265
+      0x781E0004,  //  0029  JMPF	R7	#002F
+      0x881C0D09,  //  002A  GETMBR	R7	R6	R265
+      0x781E0002,  //  002B  JMPF	R7	#002F
+      0x8C1C0D09,  //  002C  GETMET	R7	R6	R265
+      0x7C1C0200,  //  002D  CALL	R7	1
+      0x70020060,  //  002E  JMP		#0090
+      0x1C1C0300,  //  002F  EQ	R7	R1	R256
+      0x781E0004,  //  0030  JMPF	R7	#0036
+      0x881C0D00,  //  0031  GETMBR	R7	R6	R256
+      0x781E0002,  //  0032  JMPF	R7	#0036
+      0x8C1C0D00,  //  0033  GETMET	R7	R6	R256
+      0x7C1C0200,  //  0034  CALL	R7	1
+      0x70020059,  //  0035  JMP		#0090
+      0x1C1C030A,  //  0036  EQ	R7	R1	R266
+      0x781E0004,  //  0037  JMPF	R7	#003D
+      0x881C0D0A,  //  0038  GETMBR	R7	R6	R266
+      0x781E0002,  //  0039  JMPF	R7	#003D
+      0x8C1C0D0A,  //  003A  GETMET	R7	R6	R266
+      0x7C1C0200,  //  003B  CALL	R7	1
+      0x70020052,  //  003C  JMP		#0090
+      0x1C1C030B,  //  003D  EQ	R7	R1	R267
+      0x781E0004,  //  003E  JMPF	R7	#0044
+      0x881C0D0B,  //  003F  GETMBR	R7	R6	R267
+      0x781E0002,  //  0040  JMPF	R7	#0044
+      0x8C1C0D0B,  //  0041  GETMET	R7	R6	R267
+      0x7C1C0200,  //  0042  CALL	R7	1
+      0x7002004B,  //  0043  JMP		#0090
+      0x1C1C030C,  //  0044  EQ	R7	R1	R268
+      0x781E0004,  //  0045  JMPF	R7	#004B
+      0x881C0D0C,  //  0046  GETMBR	R7	R6	R268
+      0x781E0002,  //  0047  JMPF	R7	#004B
+      0x8C1C0D0C,  //  0048  GETMET	R7	R6	R268
+      0x7C1C0200,  //  0049  CALL	R7	1
+      0x70020044,  //  004A  JMP		#0090
+      0x1C1C030D,  //  004B  EQ	R7	R1	R269
+      0x781E0004,  //  004C  JMPF	R7	#0052
+      0x881C0D0D,  //  004D  GETMBR	R7	R6	R269
+      0x781E0002,  //  004E  JMPF	R7	#0052
+      0x8C1C0D0D,  //  004F  GETMET	R7	R6	R269
+      0x7C1C0200,  //  0050  CALL	R7	1
+      0x7002003D,  //  0051  JMP		#0090
+      0x1C1C030E,  //  0052  EQ	R7	R1	R270
+      0x781E0004,  //  0053  JMPF	R7	#0059
+      0x881C0D0E,  //  0054  GETMBR	R7	R6	R270
+      0x781E0002,  //  0055  JMPF	R7	#0059
+      0x8C1C0D0E,  //  0056  GETMET	R7	R6	R270
+      0x7C1C0200,  //  0057  CALL	R7	1
+      0x70020036,  //  0058  JMP		#0090
+      0x1C1C030F,  //  0059  EQ	R7	R1	R271
+      0x781E0004,  //  005A  JMPF	R7	#0060
+      0x881C0D0F,  //  005B  GETMBR	R7	R6	R271
+      0x781E0002,  //  005C  JMPF	R7	#0060
+      0x8C1C0D0F,  //  005D  GETMET	R7	R6	R271
+      0x7C1C0200,  //  005E  CALL	R7	1
+      0x7002002F,  //  005F  JMP		#0090
+      0x1C1C0310,  //  0060  EQ	R7	R1	R272
+      0x781E0004,  //  0061  JMPF	R7	#0067
+      0x881C0D10,  //  0062  GETMBR	R7	R6	R272
+      0x781E0002,  //  0063  JMPF	R7	#0067
+      0x8C1C0D10,  //  0064  GETMET	R7	R6	R272
+      0x7C1C0200,  //  0065  CALL	R7	1
+      0x70020028,  //  0066  JMP		#0090
+      0x1C1C0311,  //  0067  EQ	R7	R1	R273
+      0x781E0004,  //  0068  JMPF	R7	#006E
+      0x881C0D11,  //  0069  GETMBR	R7	R6	R273
+      0x781E0002,  //  006A  JMPF	R7	#006E
+      0x8C1C0D11,  //  006B  GETMET	R7	R6	R273
+      0x7C1C0200,  //  006C  CALL	R7	1
+      0x70020021,  //  006D  JMP		#0090
+      0x1C1C0312,  //  006E  EQ	R7	R1	R274
+      0x781E0004,  //  006F  JMPF	R7	#0075
+      0x881C0D12,  //  0070  GETMBR	R7	R6	R274
+      0x781E0002,  //  0071  JMPF	R7	#0075
+      0x8C1C0D12,  //  0072  GETMET	R7	R6	R274
+      0x7C1C0200,  //  0073  CALL	R7	1
+      0x7002001A,  //  0074  JMP		#0090
+      0x1C1C0313,  //  0075  EQ	R7	R1	R275
+      0x781E0004,  //  0076  JMPF	R7	#007C
+      0x881C0D13,  //  0077  GETMBR	R7	R6	R275
+      0x781E0002,  //  0078  JMPF	R7	#007C
+      0x8C1C0D13,  //  0079  GETMET	R7	R6	R275
+      0x7C1C0200,  //  007A  CALL	R7	1
+      0x70020013,  //  007B  JMP		#0090
+      0x1C1C0314,  //  007C  EQ	R7	R1	R276
+      0x781E0004,  //  007D  JMPF	R7	#0083
+      0x881C0D14,  //  007E  GETMBR	R7	R6	R276
+      0x781E0002,  //  007F  JMPF	R7	#0083
+      0x8C1C0D14,  //  0080  GETMET	R7	R6	R276
+      0x7C1C0200,  //  0081  CALL	R7	1
+      0x7002000C,  //  0082  JMP		#0090
+      0x1C1C0311,  //  0083  EQ	R7	R1	R273
+      0x781E0004,  //  0084  JMPF	R7	#008A
+      0x881C0D15,  //  0085  GETMBR	R7	R6	R277
+      0x781E0002,  //  0086  JMPF	R7	#008A
+      0x8C1C0D15,  //  0087  GETMET	R7	R6	R277
+      0x7C1C0200,  //  0088  CALL	R7	1
+      0x70020005,  //  0089  JMP		#0090
+      0x1C1C0315,  //  008A  EQ	R7	R1	R277
+      0x781E0003,  //  008B  JMPF	R7	#0090
+      0x881C0D15,  //  008C  GETMBR	R7	R6	R277
+      0x781E0001,  //  008D  JMPF	R7	#0090
+      0x8C1C0D15,  //  008E  GETMET	R7	R6	R277
+      0x7C1C0200,  //  008F  CALL	R7	1
+      0xA8040001,  //  0090  EXBLK	1	1
+      0x7002000B,  //  0091  JMP		#009E
+      0xAC1C0002,  //  0092  CATCH	R7	0	2
+      0x70020008,  //  0093  JMP		#009D
+      0xA4262C00,  //  0094  IMPORT	R9	R278
+      0x6028000F,  //  0095  GETGBL	R10	G15
+      0x8C2C1317,  //  0096  GETMET	R11	R9	R279
+      0x58340018,  //  0097  LDCONST	R13	K24
+      0x5C380E00,  //  0098  MOVE	R14	R7
+      0x5C3C1000,  //  0099  MOVE	R15	R8
+      0x7C2C0800,  //  009A  CALL	R11	4
+      0x7C280200,  //  009B  CALL	R10	1
+      0x70020000,  //  009C  JMP		#009E
+      0xB0080000,  //  009D  RAISE	2	R0	R0
+      0x7001FF85,  //  009E  JMP		#0025
+      0x58140019,  //  009F  LDCONST	R5	K25
+      0xAC140200,  //  00A0  CATCH	R5	1	0
+      0xB0080000,  //  00A1  RAISE	2	R0	R0
+      0x80000000,  //  00A2  RET	0	R0
     })
   )
 );